flock the crontabs so they won't overlap. Split the sge test per command. Generate the job file before qsub is called.
