+++ forecasters.js/qpf/templates/job.mustache
@@ -1,6 +1,6 @@
 #!/bin/bash
-#$ -pe 16cpn 16
-##$ -pe smp 4
+##$ -pe 16cpn 16
+#$ -pe smp 4
 #$ -N {{name}}
 #$ -l h_rt=00:15:00
