+++ forecasters.js/qpf/test/sge_qhold.js
@@ -1,0 +1,8 @@
+var debug = require('debug')('sge');
+
+var sge = require('../sge.js');
+
+sge.qhold('IFC_QPE')
+.catch(function (err) {
+  return debug(err);
+});
